+++ semantic-conventions/src/opentelemetry/semconv/model/semantic_convention.py
@@ -198,10 +198,6 @@ class ResourceSemanticConvention(BaseSemanticConvention):
     GROUP_TYPE_NAME = "resource"
 
 
-class ScopeSemanticConvention(BaseSemanticConvention):
-    GROUP_TYPE_NAME = "scope"
-
-
 class AttributeGroupConvention(BaseSemanticConvention):
     GROUP_TYPE_NAME = "attribute_group"
 
@@ -239,26 +235,7 @@ class EventSemanticConvention(BaseSemanticConvention):
             )
 
 
-class UnitSemanticConvention(BaseSemanticConvention):
-    GROUP_TYPE_NAME = "units"
-
-    allowed_keys = (  # We completely override base semantic keys here.
-        "id",
-        "type",
-        "brief",
-        "members",
-    )
-
-    def __init__(self, group, validation_ctx):
-        super().__init__(group, validation_ctx)
-        self.members = UnitMember.parse(group.get("members"), validation_ctx)
-
-
-class MetricGroupSemanticConvention(BaseSemanticConvention):
-    GROUP_TYPE_NAME = "metric_group"
-
-
-class MetricSemanticConvention(MetricGroupSemanticConvention):
+class MetricSemanticConvention(BaseSemanticConvention):
     GROUP_TYPE_NAME = "metric"
 
     allowed_keys: Tuple[str, ...] = BaseSemanticConvention.allowed_keys + (
@@ -600,10 +577,7 @@ CONVENTION_CLS_BY_GROUP_TYPE = {
         SpanSemanticConvention,
         ResourceSemanticConvention,
         EventSemanticConvention,
-        MetricGroupSemanticConvention,
         MetricSemanticConvention,
-        UnitSemanticConvention,
-        ScopeSemanticConvention,
         AttributeGroupConvention,
     )
 }